Fixed an Error with the initialization of the correlation values.
Changed the trainers to have the prime offset instead of the activation functions.

Made it possible to change the connections initial weights using a scale and offset.

Added a few more Utility Functions.

Added a funtion to make sure the CCTrainer doesn't stagnate when not using validation.

Added the ability to reset weight values in CC if they grow too large.