Fixed the way scale and offset are used for network.
Removed it from the AddConnection function and created network specific Connect Functions that are used to build the networks instead of the static functions from the connection class.

This is more elegant than the way I was doing it before and doesn't interfere with the way AddConnection was suposed to be used.