Adding the reverberating technique to the CC network trainer.
This is adding a subclass to the trainer and the CCNetwork while trying to reuse as much code as possible. I've also made some housekeeping changes to other classes, and added some functions to help generalize the existing classes to the new method.

Some of the things I've done have been commented as hacky so I know to go back and change them to a better implementation when I have time.