Added convience functions
Added some comments and added a function that will let me activate the
primary network of the pair from the trainer and get the correct outputs
back. This function as been added to the python library as well.

Down the line it might be worthwhile to rework the Rev stuff so it acts
more like the other trianers. Meaning the RevCCNetwork actually contains
the pair, and the current revccnetwork gets renamed to autoassocitive
network to more accurately reflect what it does.

Not sure how long that would take so I'm putting it off until we get
the first journal out the door.