Added functions to rename the network components.
This is necessary for when I need to modify the network structure through
insertion. Meaning nodes are added in between others instead of being
added to the end.

To do this all the nodes on the layer that come after the inserted node
are renamed and their connections are renamed.

The rename function belongs to the network at the top layer because all of
the connections that were changed need to have their keys updated in the
network's map.