Now can add new inputs from the CCTrainer.
Can now add new inputs from the CCtrainer. The indexes of the new inputs
are used to create new inputs and a new dataset with the new input vectors
is set. Right now this just adds the inputs without connecting to existing
hidden units. Need to add that ability next.