finished up the insert new inputs
All of the code should be in place to add new inputs either by just adding
and building new struture or by connecting the new inputs to the existing
hidden layer and allowing the network to update those weights.

The insert inputs function will have to be called then a new round of
training to convergence should be called.

This is only added to the CCnetwork and trainer.