Added a flag to skip shuffling.
This is a partial solution to keep the dataset from getting shuffled.

This might be important if the networks need to have the recursion added.
The order of the data could be important.

If this is true I'll also need to goback and rewrite the fillsubset
methods to ignore the distribution of the outclasses.