Changing the way Observational data is parsed
I'm making it so that instead of going from 0-1 only. I'm splitting the
outputs up so that the negative directions have their own output.
Hopefully this will make it easier for the network to learn the problem.

Especially for Cong and Yinjie's data. There is an issue though in the way
the outputs are concatinated. It's giving a false size in the C code.