Fixed the error with the ds size
The error was caused by numpy concating the arrays using slices. It caused
the size as reported by pyublas to be incorrect.

There was also an issue from the the realOuts fix. I was checking for it
even when testwhile train wasn't turned on meaning I was looking for it
when the ds wasn't there. Pointer error.