Turned off Caching
Caching seemed to be causing a problem while trying to learn the band
datasets. It seems like there is an error in the way i'm trying to cache
that is causeing the CC not to learn. When turned off the training works
again.

This pretty much invalidates all the experiments I've done with the Obs
Learning Stuff.