Added validation to update net
Added the validation to the updatenet function if the dataset has a
validation set.

Also made some changes so that the clean reverbs part of RevCC only
thresholds the outputs. The contexts are auto generated if the
numContext variable is set above 0.

Also added the validation to the revcctrainer trainouts function.