NetworkNode Mostly Complete
I believe the network node class is mostly done.

now I need to finish up the kbccnetwork class and the trainer.

kbccnetwork will keep a list of the subnetworks to make serialization
easier.

kbcctrainer will be pretty much the same with the addition of the
subnetwork candidate list and the changes to the candidate training
phase.

Serializing the kbccnetwork will need to be handled differently just
like the other networks, in that it will have a subnetwork of either
ccnetwork or kbccnetwork.